Avoids hanging after TimoutError from TCP socket
Every once in a while the code would detect a TimeoutError whilst
waiting for output from the tcp socket. This would close the process
of handling the connection, and not building it up again.

The change now is to continue the loop and await again for input
from the tcp socket.